requirements.txt
# python-telegram-bot==20.7
# yfinance==0.2.18
# pandas==2.0.3
# pytz==2023.3
# asyncio

python-telegram-bot==20.7
yfinance==0.2.28
pandas==2.2.0
numpy==1.26.4
pytz==2024.1
asyncio